waed

+++ Saraha.Core/Data/Postlike.cs
@@ -8,7 +8,7 @@ namespace Saraha.Core.Data
     public partial class Postlike
     {
       
-        public decimal LikeId { get; set; }
+        public int LikeId { get; set; }
 
         public DateTime? LikeDate { get; set; }
         public decimal? UserId { get; set; }
